Complete ggplot2 Theme for Non-Data Aesthetics

Custom alternative to the ggtheme options built into ggplot2. Removes gray boxes and grid lines from plot background. Increases font size of tick marks and axis labels. Removes gray box from legend background and legend key. Removes legend title.

Usage

theme_lyon(title_size = 16, text_size = 13)

Arguments

title_size

(numeric) size of font in axis titles

text_size

(numeric) size of font in tick labels

Value

(ggplot theme) list of ggplot2 theme elements